Address missing elements in course schedule
The upcoming edition of Joy of Cryptography for some reason nuked the
chapters on Message Authentication Codes and Block Cipher Modes of
Operation. I'm creating a session for those topics right after the
collision-resistant hash functions session. We can use the chapters from
the previous Joy of Cryptography edition, but frankly I think I'll find
something else for that session.